build: v1.6.0-rc1 (#198)
* upgrade venus

* upgrade gateway

* feat: load builtin actors from bundle

* feat: use ~/.venus-messager store files

* chore: upgrade venus

* fix: Simplify the configuration

* cmd: hidden --config flag

* chore: upgrade venus & gateway & auth

* chore: Remove the dependency on golang.org/x/xerrors

* build: bump version to v1.6.0-rc1

* cmd: rmeove --sqlite-file flag

* fix: remove messager.toml

* doc: update README.md

* cmd: Fix incorrect configuration loading

* cmd: fix actor code not registed

* chore: upgrade gateway & venus

* chore: upgrade venus & venus-auth

* doc: update READMD.md
