+++ config/modules.go
@@ -10,7 +10,7 @@ import (
 func ConfigExist(path string) (bool, error) {
 	_, err := os.Stat(path)
 	if err != nil {
-		if !os.IsExist(err) {
+		if os.IsNotExist(err) {
 			return false, nil
 		}
 		return false, err
