+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"context"
 	"encoding/hex"
 	"errors"
 	"fmt"
@@ -11,10 +10,7 @@ import (
 	"path/filepath"
 
 	"github.com/filecoin-project/venus-messager/metrics"
-	"github.com/filecoin-project/venus/pkg/util/blockstoreutil"
 	v1 "github.com/filecoin-project/venus/venus-shared/api/chain/v1"
-	builtinactors "github.com/filecoin-project/venus/venus-shared/builtin-actors"
-	"github.com/filecoin-project/venus/venus-shared/types"
 	"github.com/mitchellh/go-homedir"
 
 	ma "github.com/multiformats/go-multiaddr"
@@ -94,10 +90,6 @@ var runCmd = &cli.Command{
 		&cli.StringFlag{
 			Name:  "db-type",
 			Usage: "which db to use. sqlite/mysql",
-		},
-		&cli.StringFlag{
-			Name:  "sqlite-file",
-			Usage: "the path and file name of SQLite, eg. ~/sqlite/message.db",
 		},
 		&cli.StringFlag{
 			Name:  "mysql-dsn",
@@ -196,7 +188,7 @@ func runAction(ctx *cli.Context) error {
 		return err
 	}
 
-	if err := loadBuiltinActors(ctx.Context, repoPath, cfg); err != nil {
+	if err := ccli.LoadBuiltinActors(ctx.Context, cfg); err != nil {
 		return err
 	}
 
@@ -332,9 +324,6 @@ func updateFlag(cfg *config.Config, ctx *cli.Context) error {
 		cfg.DB.Type = ctx.String("db-type")
 		switch cfg.DB.Type {
 		case "sqlite":
-			if ctx.IsSet("sqlite-file") {
-				cfg.DB.Sqlite.File = ctx.String("sqlite-file")
-			}
 		case "mysql":
 			if ctx.IsSet("mysql-dsn") {
 				cfg.DB.MySql.ConnectionString = ctx.String("mysql-dsn")
@@ -368,41 +357,6 @@ func genSecret(cfg *config.JWTConfig) error {
 	}
 
 	return nil
-}
-
-func loadBuiltinActors(ctx context.Context, repoPath string, cfg *config.Config) error {
-	full, closer, err := service.NewNodeClient(ctx, &cfg.Node)
-	if err != nil {
-		return err
-	}
-	defer closer()
-	networkName, err := full.StateNetworkName(ctx)
-	if err != nil {
-		return err
-	}
-	builtinactors.SetNetworkBundle(networkNameToNetworkType(networkName))
-	if err := os.Setenv(builtinactors.RepoPath, repoPath); err != nil {
-		return fmt.Errorf("failed to set env %s", builtinactors.RepoPath)
-	}
-
-	bs := blockstoreutil.NewMemory()
-
-	return builtinactors.FetchAndLoadBundles(ctx, bs, builtinactors.BuiltinActorReleases)
-}
-
-func networkNameToNetworkType(networkName types.NetworkName) types.NetworkType {
-	switch networkName {
-	case "mainnet":
-		return types.NetworkMainnet
-	case "calibrationnet":
-		return types.NetworkCalibnet
-	case "butterflynet":
-		return types.NetworkButterfly
-	case "interopnet":
-		return types.NetworkInterop
-	default:
-		return types.Network2k
-	}
 }
 
 func hasFSRepo(repoPath string) (bool, error) {
