--- messager.toml
@@ -1,61 +1,0 @@
-
-[api]
-  Address = "/ip4/0.0.0.0/tcp/39812"
-
-[db]
-  type = "sqlite"
-
-  [db.mysql]
-    connMaxLifeTime = "1m0s"
-    connectionString = ""
-    debug = false
-    maxIdleConn = 10
-    maxOpenConn = 10
-
-  [db.sqlite]
-    debug = false
-    file = "./message.db"
-
-[gateway]
-  remoteEnable = true
-  token = ""
-  url = ["/ip4/127.0.0.1/tcp/45132"]
-
-  [gateway.cfg]
-    RequestQueueSize = 30
-    RequestTimeout = "5m0s"
-
-[jwt]
-  authURL = "http://127.0.0.1:8989"
-
-  [jwt.local]
-    secret = "b0b574ef3a1433df96dd0696879df4b273631e4620a56c7ab990ac6e3874c319"
-    token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoidmVudXMtbWVzc2FnZXItYWRtaW4iLCJwZXJtIjoiYWRtaW4iLCJleHQiOiIifQ.sn5UvOXIITenKt9_hnMiXY0Ib9-aLmxt8itIx4KnmIM"
-
-[log]
-  level = "info"
-  path = "messager.log"
-
-[messageService]
-  WaitingChainHeadStableDuration = "8s"
-  skipProcessHead = false
-  skipPushMessage = false
-  tipsetFilePath = "./tipset.json"
-
-[messageState]
-  CleanupInterval = 86400
-  DefaultExpiration = 259200
-  backTime = 86400
-
-[node]
-  token = ""
-  url = "/ip4/127.0.0.1/tcp/3453"
-
-[rateLimit]
-  redis = ""
-
-[tracing]
-  JaegerEndpoint = ""
-  JaegerTracingEnabled = false
-  ProbabilitySampler = 1.0
-  ServerName = "venus-messenger"
